Move ParserContext to script::parse
